scripts/requirements.txt
anthropic
psycopg2-binary
python-dotenv
